Rework Dockerfile, general improvements
This reworks the Dockerfile for a few improvements:

- uses Python 3 (thanks #33)
- uses conda for packages where possible
- uses nodejs 6.x instead of 14.04's 0.10.x
- uses phusion baseimage's daemons to our advantage
- makes more logs available to `docker logs`

This makes additional changes to the code itself:

Remove dependency on CWD. This means you can run this without a shell or, more
importantly, without being in the project directory. This is done by adding a
`APP_ROOT` constant to the cchecker_web module.

This also removes some dependencies that are not used directly. This makes for
easier migration to conda. Also bumps some dependency versions where
appropriate.

Some limitations:

cc-plugin-ncei cc-plugin-glider are not included because they haven't been
updated to support Python 3 yet (see ioos/cc-plugin-ncei#15 and
ioos/cc-plugin-glider#7)