Remove function that alters the dimensions of .cdl
Altering the dimensions of the user-supplied .cdl file causes some checks
to fail where they would otherwise pass and generates inconcistencies when
testing against a .cdl versus a .nc(4) file. Remove the function which
changes the dimensions to `1`.

The `my_init.d/run.sh` script generates a new config.yml each time it is
run, which is every time a new Docker image is built. A config.yml is already
included in the version control; changing this each time may lead to confuse
developers who choose to put certain options in to the config.yml and having
it change without them knowing.

Update the image the docker-compose.yml uses to "ioos/compliance-checker-web"
instead of the colloquial and unofficial "test_ccweb".